last work done

+++ libs/General/Headers/String.h
@@ -10,18 +10,18 @@
 
 typedef struct {
     char* data;
-    uint64_t size;
-    uint64_t allocated;
+    size_t size;
+    size_t allocated;
 } String;
 
 String*  string_init     (const char* data);
 String*  string_copy     (String* str);
-String*  string_reserve  (String* str, uint64_t to_size);
+String*  string_reserve  (String* str, size_t to_size);
 void     string_delete   (String* str);
 String*  string_append   (String* str, const char c);
 String*  string_prepend  (String* str, const char c);
 String*  string_insert   (String* str, const char* data);
-String*  string_insertAt (String* str,uint64_t index, const char* data);
+String*  string_insertAt (String* str, size_t index, const char* data);
 String*  string_popBack  (String* str);
 String*  string_popFront (String* str);
 char*    string_get      (String* str);
