+++ libs/General/Source/String.c
@@ -17,7 +17,7 @@ String*  string_copy(String* str){
     return string_init(str->data);
 }
 
-String* string_reserve(String* str, uint64_t to_size){
+String* string_reserve(String* str, size_t to_size){
     if(to_size <= str->allocated)return str;
 
     str->allocated = to_size;
@@ -45,7 +45,7 @@ String* string_prepend(String* str, const char c){
         str->allocated += SIZE_INCREMENTATION_CONSTANT;
         str->data = (char*)realloc(str->data,str->allocated);
     }
-    for(uint64_t i = str->size++;0 < i;--i){
+    for(size_t i = str->size++;0 < i;--i){
         str->data[i] = str->data[i-1];
     }
     str->data[0] = c;
@@ -53,21 +53,21 @@ String* string_prepend(String* str, const char c){
 }
 
 String* string_insert(String* str, const char* data){
-    const uint64_t len = strlen(data);
+    const size_t len = strlen(data);
     if(str->allocated - str->size < len){
         str->allocated = str->size + len;
         str->data = (char*)realloc(str->data,str->allocated);
     }
 
-    uint64_t i = 0;
+    size_t i = 0;
     while(i < len){
         str->data[str->size++] = data[i++];
     }
     return str;
 }
 
-String* string_insertAt(String* str, uint64_t index, const char* data){
-    const uint64_t len = strlen(data);
+String* string_insertAt(String* str, size_t index, const char* data){
+    const size_t len = strlen(data);
     if(str->allocated - str->size < len){
         str->allocated = str->size + len;
         str->data = (char*)realloc(str->data,str->allocated);
@@ -93,7 +93,7 @@ String* string_popBack(String* str){
 
 String* string_popFront(String* str){
     if(str->size != 0){
-        for(uint64_t i = 1;i < str->size;++i){
+        for(size_t i = 1;i < str->size;++i){
             str->data[i - 1] = str->data[i];
         }
         --str->size;
@@ -121,7 +121,7 @@ String* string_reverse(String* str){
 String* string_filter (String* str, bool(*predicate)(const char)){
     String* result = string_init(NULL);
 
-    for(uint64_t i = 0;i < str->size; ++i){
+    for(size_t i = 0;i < str->size; ++i){
         if(predicate(str->data[i])){
             string_append(result,str->data[i]);
         }
